big update
porting in html and scss from the sassessentials course.  why the hell it had to be so complex using gulp in that course made no sense to me.  using instructions from https://webdesign.tutsplus.com/tutorials/watch-and-compile-sass-in-five-quick-steps--cms-28275, and just running a simple node-sass watch is actually working.  need to see if it will work on my work computer.  will i need to install node sass dependency on that computer, or is it now part of the project already?